Delete on cascade, separating form sending, bugs in location adding
Delete on cascade wasn’t set up in the database and foreign keys
weren’t turned on! They are now. Made separate buttons for previewing
and sending the 2 different sets of forms. Fixed a bug where an edited
location would be overwritten by a newly added location. Enforcing
location names as unique at the top level because they have to be
unique in the database.